fix : bug fix of NPE
