+++ src/test/java/nextstep/subway/maps/map/application/PathServiceTest.java
@@ -3,16 +3,21 @@ package nextstep.subway.maps.map.application;
 import com.google.common.collect.Lists;
 import nextstep.subway.maps.line.domain.Line;
 import nextstep.subway.maps.line.domain.LineStation;
+import nextstep.subway.maps.map.domain.PathType;
 import nextstep.subway.maps.map.domain.SubwayPath;
 import nextstep.subway.maps.station.domain.Station;
 import nextstep.subway.utils.TestObjectUtils;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -28,7 +33,7 @@ class PathServiceTest {
         stations.put(3L, TestObjectUtils.createStation(3L, "양재역"));
         stations.put(4L, TestObjectUtils.createStation(4L, "남부터미널역"));
 
-        Line line1 = TestObjectUtils.createLine(1L, "2호선", "GREEN", 0, 10);
+        Line line1 = TestObjectUtils.createLine(1L, "2호선", "GREEN", 0, 3);
         line1.addLineStation(new LineStation(1L, null, 0, 0));
         line1.addLineStation(new LineStation(2L, 1L, 2, 2));
 
@@ -36,7 +41,7 @@ class PathServiceTest {
         line2.addLineStation(new LineStation(2L, null, 0, 0));
         line2.addLineStation(new LineStation(3L, 2L, 2, 1));
 
-        Line line3 = TestObjectUtils.createLine(3L, "3호선", "ORANGE", 0, 10);
+        Line line3 = TestObjectUtils.createLine(3L, "3호선", "ORANGE", 0, 5);
         line3.addLineStation(new LineStation(1L, null, 0, 0));
         line3.addLineStation(new LineStation(4L, 1L, 1, 2));
         line3.addLineStation(new LineStation(3L, 4L, 2, 2));
@@ -46,17 +51,42 @@ class PathServiceTest {
         pathService = new PathService();
     }
 
-    @Test
-    void findPathByArrivalTime() {
-        //given
-        LocalDateTime departTime = LocalDateTime.of(2020, 7, 1, 6, 15);
+    @DisplayName("최단 거리, 최소 소요시간 경로 응답")
+    @ParameterizedTest(name = "{0} 에서 {1} 까지 출발했을 때 {2}에 맞춰 적절한 경로를 응답한다.")
+    @MethodSource("findPathSourceProvider")
+    void findPath(Long source, Long target, PathType type, List<Long> expectedPath) {
+        //when
+        SubwayPath path = pathService.findPath(lines, source, target, type, null);
+
+        //then
+        assertThat(path.extractStationId())
+                .containsExactlyElementsOf(expectedPath);
+    }
+
+    private static Stream<Arguments> findPathSourceProvider() {
+        return Stream.of(
+                Arguments.of(1L, 3L, PathType.DISTANCE, Lists.newArrayList(1L, 4L, 3L)),
+                Arguments.of(1L, 3L, PathType.DURATION, Lists.newArrayList(1L, 2L, 3L)),
+                Arguments.of(3L, 1L, PathType.DURATION, Lists.newArrayList(3L, 2L, 1L))
+        );
+    }
+
+    @DisplayName("가장 빠른 경로 응답")
+    @ParameterizedTest(name = "{0} 에서 {1} 까지 {3} 에 출발했을 때 가장 빠른 경로를 응답한다.")
+    @MethodSource("findPathByArrivalTimeSourceProvider")
+    void findPathByArrivalTime(Long source, Long target, List<Long> expectedPath, LocalDateTime departureTime) {
         // when
-        SubwayPath subwayPath = pathService.findPathByArrivalTime(lines, 1L, 3L, departTime);
+        SubwayPath subwayPath = pathService.findPathByArrivalTime(lines, source, target, departureTime);
 
         // then
-        assertThat(subwayPath.extractStationId().size()).isEqualTo(3);
-        assertThat(subwayPath.extractStationId().get(0)).isEqualTo(1L);
-        assertThat(subwayPath.extractStationId().get(1)).isEqualTo(4L);
-        assertThat(subwayPath.extractStationId().get(2)).isEqualTo(3L);
+        assertThat(subwayPath.extractStationId())
+                .containsExactlyElementsOf(expectedPath);
+    }
+
+    private static Stream<Arguments> findPathByArrivalTimeSourceProvider() {
+        return Stream.of(
+                Arguments.of(1L, 3L, Lists.newArrayList(1L, 4L, 3L), LocalDateTime.of(2020, 7, 1, 6, 15)),
+                Arguments.of(3L, 1L, Lists.newArrayList(3L, 4L, 1L), LocalDateTime.of(2020, 7, 1, 6, 5))
+        );
     }
 }
